requirements.txt
absl-py==0.4.1
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.6.2
anaconda-project==0.8.2
api==0.0.7
appnope==0.1.0
appscript==1.0.1
arff==0.9
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.2.1
bz2file==0.98
CacheMan==2.0.8
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
colored==1.3.93
community==1.0.0b1
contextlib2==0.5.5
coverage==4.5.1
cryptography==2.4.2
cycler==0.10.0
CyHunspell==1.2.1
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
distributed==1.21.8
Django==2.1.3
django-classy-tags==0.8.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-nyt==1.1.2
django-sekizai==0.10.0
docutils==0.14
EasyProcess==0.2.3
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
funcy==1.11
future==0.16.0
gast==0.2.0
gensim==3.4.0
gevent==1.3.0
glob2==0.6
gmpy2==2.0.8
google==2.0.1
greenlet==0.4.13
grpcio==1.12.1
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
ipdb==0.10.3
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
joblib==0.13.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
langid==1.1.6
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
mysqlclient==1.3.12
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
neo4j-driver==1.6.1
neotime==1.0.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.15.1
numpydoc==0.8.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
page==1.0.16
pandas==0.23.4
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patool==1.12
patsy==0.5.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.4.1
pkginfo==1.4.2
plotly==3.4.2
pluggy==0.6.0
ply==3.11
progress==1.3
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.5
ptyprocess==0.5.2
py==1.5.3
py4j==0.10.7
pycairo==1.18.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyemd==0.5.1
pyflakes==1.6.0
Pygments==2.2.0
pyLDAvis==2.1.2
pylint==1.8.4
pymongo==3.4.0
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pyspark==2.4.0
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-igraph==0.7.1.post6
python-Levenshtein==0.12.0
python-louvain==0.11
pytz==2018.4
pyunpack==0.1.2
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
rarfile==3.0
requests==2.18.4
retrying==1.3.3
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.0
scikit-learn==0.19.1
scikit-multilearn==0.1.0
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
snowballstemmer==1.2.1
sorl-thumbnail==12.5.0
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.2
tblib==1.3.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2+2.gc449c86
toolz==0.9.0
tornado==5.0.2
tqdm==4.25.0
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wiki==0.4.2
wikipedia==1.4.0
Wikipedia-API==0.3.7
wikitools==1.3
wordcloud==1.5.0
wrapt==1.10.11
xgboost==0.81
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.2.0
zict==0.1.3
